backend_py/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
websockets==12.0
ccxt==4.2.0
redis==5.0.1
pydantic==2.5.0
aiohttp==3.9.1
python-socketio==5.10.0
loguru==0.7.2
